feat: move station rpc into hook
